requirements.txt
nextcord
python-dotenv
canvasapi
beautifulsoup4
pytz
